requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit
pyyaml
joblib
ensure
box
